Improved scene system
